ajout d'une classe UserService avec ajout des méthode concernant la gestion des users + refactoring du code
