+++ TourGuide/src/test/java/tourGuide/TestTourGuideService.java
@@ -13,6 +13,7 @@ import rewardCentral.RewardCentral;
 import tourGuide.helper.InternalTestHelper;
 import tourGuide.service.RewardsService;
 import tourGuide.service.TourGuideService;
+import tourGuide.service.UserService;
 import tourGuide.user.User;
 import tourGuide.utils.ProximityAttraction;
 import tripPricer.Provider;
@@ -55,6 +56,7 @@ public class TestTourGuideService {
 	public void addUser() {
 		/*ARRANGE*/
 		GpsUtil gpsUtil = new GpsUtil();
+		UserService userService = new UserService();
 		RewardsService rewardsService = new RewardsService(gpsUtil, new RewardCentral());
 		InternalTestHelper.setInternalUserNumber(0);
 		TourGuideService tourGuideService = new TourGuideService(gpsUtil, new RewardCentral() ,rewardsService);
@@ -62,12 +64,12 @@ public class TestTourGuideService {
 		User user = new User(UUID.randomUUID(), "jon", "000", "jon@tourGuide.com");
 		User user2 = new User(UUID.randomUUID(), "jon2", "000", "jon2@tourGuide.com");
 
-		tourGuideService.addUser(user);
-		tourGuideService.addUser(user2);
+		userService.addUser(user);
+		userService.addUser(user2);
 		
 		/*ACT*/
-		User retrivedUser = tourGuideService.getUser(user.getUserName());
-		User retrivedUser2 = tourGuideService.getUser(user2.getUserName());
+		User retrivedUser = userService.getUser(user.getUserName());
+		User retrivedUser2 = userService.getUser(user2.getUserName());
 		
 		tourGuideService.tracker.stopTracking();
 		
@@ -80,6 +82,7 @@ public class TestTourGuideService {
 	public void getAllUsers() {
 		/*ARRANGE*/
 		GpsUtil gpsUtil = new GpsUtil();
+		UserService userService = new UserService();
 		RewardsService rewardsService = new RewardsService(gpsUtil, new RewardCentral());
 		InternalTestHelper.setInternalUserNumber(0);
 		TourGuideService tourGuideService = new TourGuideService(gpsUtil, new RewardCentral() ,rewardsService);
@@ -87,11 +90,11 @@ public class TestTourGuideService {
 		User user = new User(UUID.randomUUID(), "jon", "000", "jon@tourGuide.com");
 		User user2 = new User(UUID.randomUUID(), "jon2", "000", "jon2@tourGuide.com");
 
-		tourGuideService.addUser(user);
-		tourGuideService.addUser(user2);
+		userService.addUser(user);
+		userService.addUser(user2);
 		
 		/*ACT*/
-		List<User> allUsers = tourGuideService.getAllUsers();
+		List<User> allUsers = userService.getAllUsers();
 		tourGuideService.tracker.stopTracking();
 		
 		/*ASSERT*/
@@ -173,5 +176,4 @@ public class TestTourGuideService {
 		/*ASSERT*/
 		assertEquals(5, providers.size());
 	}
-	
 }
